+++ app/government/profiles/layout.tsx
@@ -1,0 +1,15 @@
+import { AuthGuard } from "@/components/auth-guard";
+
+export default function ProfilesLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <AuthGuard requireGovernment>
+      <main className="bg-[#eaeaea] dark:bg-[#252525] lg:min-h-[calc(100vh-65px)] min-h-[calc(100vh-56px)]">
+        {children}
+      </main>
+    </AuthGuard>
+  );
+}